Add a data aggregator bot
This bot opens a fake name generator websites and collects name, phone number and date of birth of fake people and saves it in an csv sheet. The number of people whose dat is to be collected is provided by the user in the prompt the bot provides the moment it executes.